win32 update

+++ src/main/tests/test-code.cpp
@@ -98,7 +98,7 @@ Y_UTEST(code)
     for(uint64_t i=1;i<limit_of<uint32_t>::maximum ; i += 1+alea.leq(1024) )
     {
 
-        const uint32_t s = isqrt(i);
+        const uint32_t s = uint32_t(isqrt(i));
         //std::cerr << i << " -> " << s << std::endl;
         Y_ASSERT(s*s<=i);
     }
